Rename align to translate
