+++ nltk/test/bleu.doctest
@@ -2,7 +2,7 @@
 BLEU tests
 ==========
 
->>> from nltk.align import bleu
+>>> from nltk.translate import bleu
 
 If the candidate has no alignment to any of the references, the BLEU score is 0.
