+++ nltk/test/translate.doctest
@@ -52,7 +52,7 @@ but they are easily inverted:
 We can create new alignments, but these need to be in the correct range of
 the corresponding sentences:
 
-    >>> from nltk.align import Alignment, AlignedSent
+    >>> from nltk.translate import Alignment, AlignedSent
     >>> als = AlignedSent(['Reprise', 'de', 'la', 'session'],
     ...                   ['Resumption', 'of', 'the', 'session'],
     ...                   Alignment([(0, 0), (1, 4), (2, 1), (3, 3)]))
@@ -78,7 +78,7 @@ EM for IBM Model 1
 
 Here is an example from Koehn, 2010:
 
-    >>> from nltk.align import IBMModel1
+    >>> from nltk.translate import IBMModel1
     >>> corpus = [AlignedSent(['the', 'house'], ['das', 'Haus']),
     ...           AlignedSent(['the', 'book'], ['das', 'Buch']),
     ...           AlignedSent(['a', 'book'], ['ein', 'Buch'])]
