+++ nltk/test/unit/translate/test_ibm5.py
@@ -6,11 +6,11 @@ Tests for IBM Model 5 training methods
 import unittest
 
 from collections import defaultdict
-from nltk.align import AlignedSent
-from nltk.align import IBMModel
-from nltk.align import IBMModel4
-from nltk.align import IBMModel5
-from nltk.align.ibm_model import AlignmentInfo
+from nltk.translate import AlignedSent
+from nltk.translate import IBMModel
+from nltk.translate import IBMModel4
+from nltk.translate import IBMModel5
+from nltk.translate.ibm_model import AlignmentInfo
 
 
 class TestIBMModel5(unittest.TestCase):
