+++ nltk/translate/__init__.py
@@ -11,13 +11,13 @@ Experimental features for machine translation.
 These interfaces are prone to change.
 """
 
-from nltk.align.api import AlignedSent, Alignment
-from nltk.align.ibm_model import IBMModel
-from nltk.align.ibm1 import IBMModel1
-from nltk.align.ibm2 import IBMModel2
-from nltk.align.ibm3 import IBMModel3
-from nltk.align.ibm4 import IBMModel4
-from nltk.align.ibm5 import IBMModel5
-from nltk.align.bleu_score import bleu
+from nltk.translate.api import AlignedSent, Alignment
+from nltk.translate.ibm_model import IBMModel
+from nltk.translate.ibm1 import IBMModel1
+from nltk.translate.ibm2 import IBMModel2
+from nltk.translate.ibm3 import IBMModel3
+from nltk.translate.ibm4 import IBMModel4
+from nltk.translate.ibm5 import IBMModel5
+from nltk.translate.bleu_score import bleu
 from nltk.translate.stack_decoder import StackDecoder
 from nltk.translate.phrase_table import PhraseTable
