+++ nltk/translate/api.py
@@ -20,7 +20,7 @@ class AlignedSent(object):
     Return an aligned sentence object, which encapsulates two sentences
     along with an ``Alignment`` between them.
 
-        >>> from nltk.align import AlignedSent
+        >>> from nltk.translate import AlignedSent
         >>> algnsent = AlignedSent(['klein', 'ist', 'das', 'Haus'],
         ...     ['the', 'house', 'is', 'small'], '0-2 1-3 2-1 3-0')
         >>> algnsent.words
@@ -216,7 +216,7 @@ class AlignedSent(object):
         Return an error rate between 0.0 (perfect alignment) and 1.0 (no
         alignment).
 
-            >>> from nltk.align import AlignedSent
+            >>> from nltk.translate import AlignedSent
             >>> s = AlignedSent(["the", "cat"], ["le", "chat"], [(0, 0), (1, 1)])
             >>> s.alignment_error_rate(s)
             0.0
@@ -262,7 +262,7 @@ class Alignment(frozenset):
     j-th element of s2.  Tuples are extensible (they might contain
     additional data, such as a boolean to indicate sure vs possible alignments).
 
-        >>> from nltk.align import Alignment
+        >>> from nltk.translate import Alignment
         >>> a = Alignment([(0, 0), (0, 1), (1, 2), (2, 2)])
         >>> a.invert()
         Alignment([(0, 0), (1, 0), (2, 1), (2, 2)])
