+++ nltk/translate/ibm5.py
@@ -114,12 +114,12 @@ Translation: Parameter Estimation. Computational Linguistics, 19 (2),
 from __future__ import division
 from collections import defaultdict
 from math import factorial
-from nltk.align import AlignedSent
-from nltk.align import Alignment
-from nltk.align import IBMModel
-from nltk.align import IBMModel4
-from nltk.align.ibm_model import Counts
-from nltk.align.ibm_model import longest_target_sentence_length
+from nltk.translate import AlignedSent
+from nltk.translate import Alignment
+from nltk.translate import IBMModel
+from nltk.translate import IBMModel4
+from nltk.translate.ibm_model import Counts
+from nltk.translate.ibm_model import longest_target_sentence_length
 import warnings
