+++ web/api/nltk.rst
@@ -11,14 +11,6 @@ nltk Package
     :undoc-members:
     :show-inheritance:
 
-:mod:`align` Module
--------------------
-
-.. automodule:: nltk.align
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 :mod:`collocations` Module
 --------------------------
 
@@ -87,6 +79,14 @@ nltk Package
 ---------------------
 
 .. automodule:: nltk.toolbox
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+:mod:`translate` Module
+-------------------
+
+.. automodule:: nltk.translate
     :members:
     :undoc-members:
     :show-inheritance:
